working bfs

+++ src/imgui/Point.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Point {
-    float x, y;
-    Point(float x = 0.0f, float y = 0.0f) : x(x), y(y) {}
+    int x, y;
+    Point(int x = 0.0f, int y = 0.0f) : x(x), y(y) {}
 };
